requirements.txt
python-binance
pandas
plotly
kaleido
openpyxl
telegram
python-telegram-bot